+++ srcs/ft_utils.c
@@ -1,18 +1,18 @@
 #include "fdf.h"
 
-float	ft_decimal_part(float num)
+double	ft_decimal_part(double num)
 {
 	if (num > 0.f)
 		return (num - (int)num);
 	return (num - (int)num + 1.f);
 }
 
-float	ft_reciprocal(float num)
+double	ft_reciprocal(double num)
 {
 	return (1.f - ft_decimal_part(num));
 }
 
-int	ft_lerp(int v0, int v1, float t)
+int	ft_lerp(int v0, int v1, double t)
 {
 	return ((int)((double)v0 + (v1 - v0) * t));
 }
